Loader bar
Created a loader in bar format and the effect with if it were a ping pong the bar goes from one side to the other.

Created 2 props for it one is the bgBar that defined the background of the loader and the other the color and sets the background of the bar.

For the color I used the variable getColor to use as the default value and for bgColor I created a new variable getBgBar and set the default to be #CCC.